Fix cast warning

+++ src/create_card_tests.cpp
@@ -65,10 +65,10 @@ TEST_CASE("run_main", "[.]") {
     REQUIRE(!std::filesystem::exists(out_dir));
     REQUIRE(std::filesystem::create_directory(out_dir));
 
-    for (uint16_t i = 0; i < entries.size(); i++) {
+    for (int i = 0; i < entries.size(); i++) {
         auto fe = entries[i];
         kanji_data kanji_data { fe.get_kanji(),
-                                i + 1,
+                                static_cast<uint16_t>(i + 1),
                                 kanjidic2_doc,
                                 jmdict_e_doc
         };
